Fix typescript problems in async thunks
