+++ src/redux-state/slices/wallet.ts
@@ -1,7 +1,5 @@
-import { createAsyncThunk, createSlice } from "@reduxjs/toolkit"
+import { createSlice } from "@reduxjs/toolkit"
 import portrait from "shared/assets/portrait.png"
-import { resolveAddressToName } from "shared/utils"
-import { ClaimState, resetClaiming, setClaimingUser } from "./claim"
 
 export type WalletState = {
   isConnected: boolean
@@ -40,66 +38,3 @@ export const { updateConnectedWallet, updateDisconnectedWallet } =
   walletSlice.actions
 
 export default walletSlice.reducer
-
-export const fetchWalletName = createAsyncThunk(
-  "wallet/fetchWalletName",
-  async ({ address }: { address: string }, { dispatch, getState }) => {
-    const {
-      claim: { useConnectedWallet },
-    } = getState() as { claim: ClaimState }
-
-    const resolvedName = await resolveAddressToName(address)
-
-    if (resolvedName) {
-      dispatch(
-        updateConnectedWallet({
-          address,
-          name: resolvedName,
-        })
-      )
-
-      if (useConnectedWallet) {
-        dispatch(setClaimingUser({ name: resolvedName, address }))
-      }
-    }
-  }
-)
-
-export const connectWalletGlobally = createAsyncThunk(
-  "wallet/connectWalletGlobally",
-  async (
-    { address, avatar }: { address: string; avatar?: string },
-    { dispatch, getState }
-  ) => {
-    const {
-      claim: { useConnectedWallet },
-    } = getState() as { claim: ClaimState }
-
-    dispatch(
-      updateConnectedWallet({
-        address,
-        avatar,
-      })
-    )
-    await dispatch(fetchWalletName({ address }))
-
-    if (useConnectedWallet) {
-      dispatch(setClaimingUser({ address }))
-    }
-  }
-)
-
-export const disconnectWalletGlobally = createAsyncThunk(
-  "wallet/disconnectWalletGlobally",
-  async (_, { dispatch, getState }) => {
-    const {
-      claim: { useConnectedWallet },
-    } = getState() as { claim: ClaimState }
-
-    dispatch(updateDisconnectedWallet())
-
-    if (useConnectedWallet) {
-      dispatch(resetClaiming())
-    }
-  }
-)
